Provide the number of coefficients to find_power_series rather than computing them every time.
The function find_power_series now takes an integer as an argument, which determines the length of the for-loop.
Instead the number of coefficients must be computed prior to the function call.

Because this change involved major changes to almost all of cascade.c, the version number has been bumped to 0.8