Remove the polynomial from acb_ode
Because the power series expansion is not intrinsic to the differential equation, it makes little sense to store it within the acb_ode.
Therefore it has been removed from the struct and must now be provided individually. The functions have been adjusted accordingly.

This change means that an acb_ode no longer stores a complete differential equation but rather just the differential operator used to define it.

Jade was never storing a polynomial inside its jade_ode data type and is therefore not affected as much.
Just a few ccalls had to be adjusted and diffEq (for 'differential equation') has been renamed to diffOp (for 'differential operator').