Change implementation of CMake
The file CMakeLists.txt has been removed from the build directory and is now located in the root directory of the repository.
To invoke the default installation, a bash script has been provided.
However this does not allow custom flags or paths or anything of the like.

It is assumed that those people, who want to set custom CMake flags, know how to run CMake anyway and therefore don't need the script.