Minor performance improvements
An additional integer variable has been introduced in the iteration loop,
which reduces the number of arithmetic operations performed by a lot. Hence performance improves at least a little bit.

Also, previously the length of the output polynomial was asserted by setting a specific coefficient to 1.
That was a bit og a hack, and it has been replaced by the more appropriate call to acb_poly_fit_length.

Notice however that this function call does not adjust the "length" value of the polynomial and so the macro
acb_poly_get_coeff_ptr will return NULL even though an appropriate acb_t is present. However these NULLs can simply be skipped.