Abstract memory management
Previously, acb_ode_reduce altered the value of the degree field internally, but didn't realloc any memory.
This was a gold mine for memory leaks and unexpected side effects.

While there is a note in the documentation, this is still stupid and so it has been changed.
Now all the memory management is performed by the library rather than manually by the user!